requirements.txt
langchain-groq    
pydantic
pymupdf
python-dotenv
apify-client
-e .